Tests broken into appropriate targets
Those that apply to STLR in their own location, same bifrucation applied to infrastructure
