+++ Tests/STLRTests/DynamicGeneratorTest.swift
